# src/infrastructure/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Games API",
    "description": "Games API with express and mysql",
    "version": "1",
    "contact": "misaelmoreira99.sm@gmail.com"
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Games API"
    }
  ],
  "paths": {
...
    "/api/v1/games/{id}": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "Returns a game.",
        "description": "Returns a game.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "game id.",
            "type": "string",
            "example": "131db526-4599-4485-803d-cca8573693d5"
          }
        ],
        "responses": {
          "404": {
            "description": "Game not exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "200": {
            "description": "Returns a game.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResourceResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "games"
        ],
        "summary": "Deletes a game.",
        "description": "Deletes a game.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "game id.",
            "type": "string",
            "example": "131db526-4599-4485-803d-cca8573693d5"
          }
        ],
        "responses": {
          "404": {
            "description": "Game not exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "204": {
            "description": "Game deleted successfully."
          }
        }
      },
      "put": {
        "tags": [
          "games"
        ],
        "summary": "Returns a game.",
        "description": "Returns a game.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "game id.",
            "type": "string",
            "example": "131db526-4599-4485-803d-cca8573693d5"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GamePostRequest"
              }
            }
          }
        },
        "responses": {
          "404": {
            "description": "Returns genres, game, platforms, or age rating that does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "200": {
            "description": "Returns an updated game.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResourceResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GamePostRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "The Witcher 3"
          },
          "price": {
            "type": "number",
            "example": 100
          },
          "description": {
            "type": "string",
            "example": "O jogo mais premiado de uma geração agora aprimorado para a atual! Experimente The Witcher 3: Wild Hunt e suas expansões nesta coleção definitiva, com melhor desempenho, visuais aprimorados, novo conteúdo adicional, modo fotografia e muito mais!"
          },
          "ageRatingId": {
            "type": "string",
            "example": "131db526-4599-4485-803d-cca8573693d5"
          },
          "platforms": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "131db526-4599-4485-803d-cca8573693d5"
            }
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "131db526-4599-4485-803d-cca8573693d5"
            }
          }
        }
      },
      "Game": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "131db526-4599-4485-803d-cca8573693d5"
          },
          "name": {
            "type": "string",
            "example": "The Witcher 3"
          },
          "description": {
            "type": "string",
            "example": "O jogo mais premiado de uma geração agora aprimorado para a atual! Experimente The Witcher 3: Wild Hunt e suas expansões nesta coleção definitiva, com melhor desempenho, visuais aprimorados, novo conteúdo adicional, modo fotografia e muito mais!"
          },
          "price": {
            "type": "number",
            "example": 100
          },
          "releaseDate": {
            "type": "string",
            "example": "2020-05-14T00:00:00.000Z"
          },
          "ageRating": {
            "type": "object",
            "$ref": "#/components/schemas/AgeRating"
          },
          "platforms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Platform"
            }
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          }
        }
      },
      "GameResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "131db526-4599-4485-803d-cca8573693d5"
          },
          "name": {
            "type": "string",
            "example": "The Witcher 3"
          },
          "description": {
            "type": "string",
            "example": "O jogo mais premiado de uma geração agora aprimorado para a atual! Experimente The Witcher 3: Wild Hunt e suas expansões nesta coleção definitiva, com melhor desempenho, visuais aprimorados, novo conteúdo adicional, modo fotografia e muito mais!"
          },
          "price": {
            "type": "number",
            "example": 100
          },
          "releaseDate": {
            "type": "string",
            "example": "2020-05-14T00:00:00.000Z"
          },
          "ageRating": {
            "type": "object",
            "$ref": "#/components/schemas/AgeRating"
          },
          "platforms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Platform"
            }
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "_links": {
            "type": "object",
            "properties": {
              "self": {
                "type": "object",
                "properties": {
                  "href": {
                    "type": "string",
                    "example": "/api/v1/games/:id"
                  }
                }
              },
              "PUT_update_game": {
                "type": "object",
                "properties": {
                  "href": {
                    "type": "string",
                    "example": "/api/v1/genres/:id"
                  }
                }
              },
              "DELETE_delete_game": {
                "type": "object",
                "properties": {
                  "href": {
                    "type": "string",
                    "example": "/api/v1/genres/:id"
                  }
                }
              }
            }
          }
        }
      },
...
      "GameResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Game"
            }
          },
          "success": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "Você enviou gêneros duplicados."
            }
          }
        }
      },
      "GameResourceResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameResource"
            }
          },
          "success": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "Você enviou gêneros duplicados."
            }
          }
        }
      },
...
      "AgeRating": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "131db526-4599-4485-803d-cca8573693d5"
          },
          "age": {
            "type": "string",
            "example": "L"
          },
          "description": {
            "type": "string",
            "example": "Livre para todos os públicos."
          }
        }
      },
...
      "Platform": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Xbox Series S"
          },
          "id": {
            "type": "string",
            "example": "131db526-4599-4485-803d-cca8573693d5"
          }
        }
      },
...
      "Genre": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "action"
          },
          "id": {
            "type": "string",
            "example": "131db526-4599-4485-803d-cca8573693d5"
          }
        }
      },
...
    }
  }
}